requirements.txt
chainlit==2.2.0
openai
python-dotenv